test: split content management tests
