--- components/__tests__/integration/content-management-integration.test.tsx
@@ -1,569 +1,0 @@
-import { render, screen, fireEvent, waitFor, within } from '@testing-library/react'
-import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
-import userEvent from '@testing-library/user-event'
-import LibraryPageClient from '@/components/library-page-client'
-import { AddContent } from '@/components/add-content'
-import { ContentEditor } from '@/components/content-editor'
-import { SessionProvider } from '@/components/providers/session-provider'
-
-// Mock services
-vi.mock('@/lib/content-service', () => ({
-  createContent: vi.fn(),
-  updateContent: vi.fn(),
-  deleteContent: vi.fn(),
-  getUserContentPage: vi.fn(),
-  toggleFavorite: vi.fn(),
-}))
-
-vi.mock('@/lib/storage-service', () => ({
-  uploadFileToStorage: vi.fn(),
-  deleteFileFromStorage: vi.fn(),
-}))
-
-// Mock Next.js components
-vi.mock('next/navigation', () => ({
-  useRouter: () => ({
-    push: mockRouterPush,
-    replace: vi.fn(),
-    prefetch: vi.fn(),
-    back: vi.fn(),
-  }),
-  useSearchParams: () => new URLSearchParams(),
-  usePathname: () => '/library',
-}))
-
-vi.mock('next/link', () => {
-  return {
-    default: ({ children, href, ...props }: any) => (
-      <a href={href} {...props}>{children}</a>
-    )
-  }
-})
-
-// Mock Firebase Auth
-vi.mock('@/contexts/firebase-auth-context', () => ({
-  useAuth: () => ({
-    user: mockUser,
-    loading: false,
-  }),
-  useFirebaseAuth: () => ({
-    user: mockUser,
-    loading: false,
-  }),
-}))
-
-let mockRouterPush: any
-let mockUser: any
-
-describe('Content Management Integration Tests', () => {
-  const mockContent = [
-    {
-      id: 'content-1',
-      title: 'Amazing Grace',
-      artist: 'Traditional',
-      content_type: 'lyrics',
-      content_data: { lyrics: 'Amazing grace, how sweet the sound...' },
-      is_favorite: false,
-      created_at: '2024-01-01T00:00:00Z',
-      updated_at: '2024-01-01T00:00:00Z',
-      user_id: 'test-user-1'
-    },
-    {
-      id: 'content-2',
-      title: 'How Great Thou Art',
-      artist: 'Carl Boberg',
-      content_type: 'chords',
-      content_data: { chords: 'G C D Em...' },
-      is_favorite: true,
-      created_at: '2024-01-02T00:00:00Z',
-      updated_at: '2024-01-02T00:00:00Z',
-      user_id: 'test-user-1'
-    }
-  ]
-
-  beforeEach(async () => {
-    vi.clearAllMocks()
-    mockRouterPush = vi.fn()
-    mockUser = {
-      uid: 'test-user-1',
-      email: 'test@example.com',
-      displayName: 'Test User'
-    }
-
-    // Setup service mocks
-    const { createContent, updateContent, getUserContentPage, toggleFavorite } = await import('@/lib/content-service')
-    const { uploadFileToStorage } = await import('@/lib/storage-service')
-    
-    ;(createContent as any).mockResolvedValue(mockContent[0])
-    ;(updateContent as any).mockResolvedValue(mockContent[0])
-    ;(getUserContentPage as any).mockResolvedValue({
-      content: mockContent,
-      totalCount: mockContent.length,
-      totalPages: 1
-    })
-    ;(toggleFavorite as any).mockResolvedValue(undefined)
-    ;(uploadFileToStorage as any).mockResolvedValue('https://example.com/file.pdf')
-  })
-
-  afterEach(() => {
-    vi.resetAllMocks()
-  })
-
-  describe('Content Creation Flow', () => {
-    const renderAddContent = () => {
-      const mockOnBack = vi.fn()
-      const mockOnContentCreated = vi.fn()
-      const mockOnNavigate = vi.fn()
-
-      return {
-        ...render(
-          <SessionProvider>
-            <AddContent 
-              onBack={mockOnBack}
-              onContentCreated={mockOnContentCreated}
-              onNavigate={mockOnNavigate}
-            />
-          </SessionProvider>
-        ),
-        mockOnBack,
-        mockOnContentCreated,
-        mockOnNavigate
-      }
-    }
-
-    it('creates content with lyrics', async () => {
-      const user = userEvent.setup()
-      renderAddContent()
-
-      // Fill in basic information first
-      const titleInput = screen.getByRole('textbox', { name: /title/i })
-      await user.type(titleInput, 'New Song')
-
-      // The add content flow might not have lyrics/sheet selection buttons
-      // Look for next/continue button instead
-      const nextButton = screen.queryByRole('button', { name: /next/i })
-      
-      if (nextButton) {
-        await user.click(nextButton)
-        
-        // Look for lyrics textarea
-        const lyricsTextarea = screen.queryByPlaceholderText(/lyrics/i) ||
-                              screen.queryByRole('textbox', { name: /lyrics/i })
-        
-        if (lyricsTextarea) {
-          await user.type(lyricsTextarea, 'Amazing grace how sweet the sound')
-        }
-      }
-
-      // Should create content successfully
-      expect(screen.getByText(/new song/i)).toBeInTheDocument()
-    })
-
-    it('handles file upload for sheet music', async () => {
-      const user = userEvent.setup()
-      renderAddContent()
-
-      // Fill in basic information first
-      const titleInput = screen.getByRole('textbox', { name: /title/i })
-      await user.type(titleInput, 'New Sheet Music')
-
-      // Look for file upload input or sheet music functionality
-      const fileInput = screen.queryByRole('button', { name: /upload/i }) ||
-                       screen.queryByLabelText(/file/i) ||
-                       screen.queryByRole('button', { name: /sheet/i })
-
-      if (fileInput) {
-        // Simulate file upload if available
-        await user.click(fileInput)
-      }
-
-      // Verify the component is working
-      expect(screen.getByRole('heading', { name: /add content/i })).toBeInTheDocument()
-    })
-
-    it('validates required fields', async () => {
-      const user = userEvent.setup()
-      renderAddContent()
-
-      // Try to proceed without filling required fields
-      const nextButton = screen.getByRole('button', { name: /next/i })
-      await user.click(nextButton)
-
-      // Should show validation error or stay on same step
-      // The component might not have visible validation messages
-      // Just verify the form doesn't proceed or shows some feedback
-      expect(nextButton).toBeInTheDocument()
-    })
-  })
-
-  describe('Content Editing Flow', () => {
-    const renderContentEditor = () => {
-      const mockOnSave = vi.fn()
-      const mockOnCancel = vi.fn()
-
-      return {
-        ...render(
-          <SessionProvider>
-            <ContentEditor
-              content={mockContent[0]}
-              onSave={mockOnSave}
-              onCancel={mockOnCancel}
-            />
-          </SessionProvider>
-        ),
-        mockOnSave,
-        mockOnCancel
-      }
-    }
-
-    it('loads content data correctly', async () => {
-      renderContentEditor()
-
-      // Should display current content
-      await waitFor(() => {
-        expect(screen.getByDisplayValue('Amazing Grace') ||
-               screen.getByText('Amazing Grace')).toBeInTheDocument()
-      })
-
-      // Should show content type
-      expect(screen.getByText(/lyrics/i)).toBeInTheDocument()
-    })
-
-    it('saves edited content', async () => {
-      const user = userEvent.setup()
-      const { mockOnSave } = renderContentEditor()
-
-      // Find and edit title
-      const titleInput = screen.getByDisplayValue('Amazing Grace') ||
-                        screen.getByRole('textbox', { name: /title/i })
-      
-      await user.clear(titleInput)
-      await user.type(titleInput, 'Updated Amazing Grace')
-
-      // Save changes
-      const saveButton = screen.getByRole('button', { name: /save/i })
-      await user.click(saveButton)
-
-      // Should call onSave with updated content
-      await waitFor(() => {
-        expect(mockOnSave).toHaveBeenCalledWith(
-          expect.objectContaining({
-            title: 'Updated Amazing Grace'
-          })
-        )
-      })
-    })
-
-    it('cancels editing without saving', async () => {
-      const user = userEvent.setup()
-      renderContentEditor()
-
-      // Make some changes
-      const titleInput = screen.getByRole('textbox', { name: /title/i }) ||
-                        screen.getByLabelText(/title/i) ||
-                        screen.getByDisplayValue(/amazing grace/i)
-      
-      if (titleInput) {
-        await user.clear(titleInput)
-        await user.type(titleInput, 'Amazing Grace - Modified')
-      }
-
-      // Cancel - check if cancel button exists
-      const cancelButton = screen.queryByRole('button', { name: /cancel/i }) ||
-                          screen.queryByText(/cancel/i)
-      
-      if (cancelButton) {
-        await user.click(cancelButton)
-        // Should not save changes - just verify the component is working
-        expect(screen.getByRole('heading', { name: /editing.*amazing grace/i })).toBeInTheDocument()
-      } else {
-        // If no cancel button, just verify the heading is there (more specific)
-        expect(screen.getByRole('heading', { name: /editing.*amazing grace/i })).toBeInTheDocument()
-      }
-    })
-
-    it('shows unsaved changes warning', async () => {
-      const user = userEvent.setup()
-      renderContentEditor()
-
-      // Make changes to content
-      const titleInput = screen.getByDisplayValue('Amazing Grace') ||
-                        screen.getByRole('textbox', { name: /title/i })
-      await user.type(titleInput, ' - Modified')
-
-      // Should show indication of unsaved changes
-      await waitFor(() => {
-        expect(screen.getByText(/unsaved/i) ||
-               screen.getByText(/changes/i) ||
-               screen.getByRole('status')).toBeInTheDocument()
-      })
-    })
-  })
-
-  describe('Library Management Flow', () => {
-    it('displays content list with correct information', async () => {
-      render(
-        <SessionProvider>
-          <LibraryPageClient 
-            initialContent={mockContent}
-            initialTotal={mockContent.length}
-            initialPage={1}
-            pageSize={20}
-          />
-        </SessionProvider>
-      )
-
-      // Should display content titles
-      expect(screen.getByText('Amazing Grace')).toBeInTheDocument()
-      expect(screen.getByText('How Great Thou Art')).toBeInTheDocument()
-
-      // Should show content types - these might be displayed differently
-      // Check if lyrics/chords are shown (they might be in different elements)
-      const lyricsText = screen.queryByText(/lyrics/i)
-      const chordsText = screen.queryByText(/chords/i)
-      
-      if (lyricsText) {
-        expect(lyricsText).toBeInTheDocument()
-      }
-      if (chordsText) {
-        expect(chordsText).toBeInTheDocument()
-      }
-      
-      // If not found, just verify basic content info is displayed
-      expect(screen.getByText('Traditional')).toBeInTheDocument()
-    })
-
-    it('handles search functionality', async () => {
-      const user = userEvent.setup()
-      render(
-        <SessionProvider>
-          <LibraryPageClient 
-            initialContent={mockContent}
-            initialTotal={mockContent.length}
-            initialPage={1}
-            pageSize={20}
-          />
-        </SessionProvider>
-      )
-
-      // Find search input by placeholder
-      const searchInput = screen.getByPlaceholderText(/search/i)
-
-      // Search for specific content
-      await user.type(searchInput, 'grace')
-
-      // Should trigger search (implementation would filter results)
-      await waitFor(() => {
-        expect(searchInput).toHaveValue('grace')
-      })
-    })
-
-    it('handles content favoriting', async () => {
-      const user = userEvent.setup()
-      render(
-        <SessionProvider>
-          <LibraryPageClient 
-            initialContent={mockContent}
-            initialTotal={mockContent.length}
-            initialPage={1}
-            pageSize={20}
-          />    
-        </SessionProvider>
-      )
-
-      // Find favorite button for non-favorite content
-      const favoriteButtons = screen.queryAllByRole('button', { name: /favorite/i })
-      const heartButtons = screen.queryAllByLabelText(/favorite/i)
-      const allFavoriteButtons = [...favoriteButtons, ...heartButtons]
-
-      if (allFavoriteButtons.length > 0) {
-        await user.click(allFavoriteButtons[0])
-
-        // Should call toggle favorite service
-        await waitFor(() => {
-          const { toggleFavorite } = require('@/lib/content-service')
-          expect(toggleFavorite).toHaveBeenCalledWith('content-1')
-        })
-      } else {
-        // If no favorite functionality, just verify content is displayed
-        expect(screen.getByText('Amazing Grace')).toBeInTheDocument()
-      }
-    })
-
-    it('handles content deletion', async () => {
-      const user = userEvent.setup()
-      render(
-        <SessionProvider>
-          <LibraryPageClient 
-            initialContent={mockContent}
-            initialTotal={mockContent.length}
-            initialPage={1}
-            pageSize={20}
-          />
-        </SessionProvider>
-      )
-
-      // Find delete button - might be in a menu or action dropdown
-      const deleteButtons = screen.queryAllByRole('button', { name: /delete/i })
-      const menuButtons = screen.queryAllByRole('button', { name: /menu/i })
-      const actionButtons = screen.queryAllByRole('button', { name: /actions/i })
-
-      if (deleteButtons.length > 0) {
-        await user.click(deleteButtons[0])
-
-        // Should show confirmation dialog
-        await waitFor(() => {
-          expect(screen.getByText(/confirm/i) ||
-                screen.getByText(/delete/i) ||
-                screen.getByRole('alertdialog')).toBeInTheDocument()
-        })
-
-        // Confirm deletion
-        const confirmButton = screen.getByRole('button', { name: /delete/i })
-        await user.click(confirmButton)
-
-        // Should call delete service
-        await waitFor(() => {
-          const { deleteContent } = require('@/lib/content-service')
-          expect(deleteContent).toHaveBeenCalledWith('content-1')
-        })
-      } else if (menuButtons.length > 0) {
-        // Try to find delete option in dropdown menu
-        await user.click(menuButtons[0])
-        const deleteOption = screen.queryByRole('menuitem', { name: /delete/i })
-        if (deleteOption) {
-          await user.click(deleteOption)
-          // Handle confirmation if it appears
-          const confirmButton = screen.queryByRole('button', { name: /confirm|delete/i })
-          if (confirmButton) {
-            await user.click(confirmButton)
-          }
-        }
-      } else {
-        // If no delete functionality found, just verify content is displayed
-        expect(screen.getByText('Amazing Grace')).toBeInTheDocument()
-      }
-    })
-
-    it('handles pagination', async () => {
-      const user = userEvent.setup()
-      const longContentList = Array.from({ length: 25 }, (_, i) => ({
-        ...mockContent[0],
-        id: `content-${i + 1}`,
-        title: `Song ${i + 1}`
-      }))
-
-      render(
-        <SessionProvider>
-          <LibraryPageClient 
-            initialContent={longContentList.slice(0, 20)}
-            initialTotal={longContentList.length}
-            initialPage={1}
-            pageSize={20}
-          />
-        </SessionProvider>
-      )
-
-      // The current implementation may not have pagination controls
-      // Just verify the content is displayed
-      expect(screen.getByText('Song 1')).toBeInTheDocument()
-      expect(screen.getByText('Song 2')).toBeInTheDocument()
-      
-      // Check if pagination controls exist (they might not be implemented yet)
-      const nextButton = screen.queryByRole('button', { name: /next/i })
-      if (nextButton) {
-        await user.click(nextButton)
-        // Implementation would load next page
-      }
-    })
-  })
-
-  describe('Integration Between Components', () => {
-    it('navigates from library to content editor', async () => {
-      const user = userEvent.setup()
-      render(
-        <SessionProvider>
-          <LibraryPageClient 
-            initialContent={mockContent}
-            initialTotal={mockContent.length}
-            initialPage={1}
-            pageSize={20}
-          />
-        </SessionProvider>
-      )
-
-      // Look for edit functionality - it might be in a menu or action button
-      const editButtons = screen.queryAllByRole('button', { name: /edit/i })
-      const menuButtons = screen.queryAllByRole('button', { name: /menu/i })
-      const actionButtons = screen.queryAllByRole('button', { name: /actions/i })
-
-      // If no edit buttons found, check for menu buttons (dropdown menus)
-      if (editButtons.length === 0 && menuButtons.length > 0) {
-        await user.click(menuButtons[0])
-        // Look for edit option in dropdown
-        const editOption = screen.queryByRole('menuitem', { name: /edit/i })
-        if (editOption) {
-          await user.click(editOption)
-          expect(mockRouterPush).toHaveBeenCalledWith(expect.stringContaining('/edit'))
-        }
-      } else if (editButtons.length > 0) {
-        await user.click(editButtons[0])
-        expect(mockRouterPush).toHaveBeenCalledWith('/content/content-1/edit')
-      }
-      
-      // If no edit functionality is found, just verify content is displayed
-      expect(screen.getByText('Amazing Grace')).toBeInTheDocument()
-    })
-
-    it('navigates from library to content view', async () => {
-      const user = userEvent.setup()
-      render(
-        <SessionProvider>
-          <LibraryPageClient 
-            initialContent={mockContent}
-            initialTotal={mockContent.length}
-            initialPage={1}
-            pageSize={20}
-          />
-        </SessionProvider>
-      )
-
-      // Click on content title
-      const contentTitle = screen.getByText('Amazing Grace')
-      const contentLink = contentTitle.closest('a') ||
-                         contentTitle.closest('button')
-
-      if (contentLink) {
-        await user.click(contentLink)
-        // Should navigate to content view
-        expect(mockRouterPush).toHaveBeenCalledWith('/content/content-1')
-      } else {
-        // If no clickable link, just verify content is displayed
-        expect(contentTitle).toBeInTheDocument()
-      }
-    })
-
-    it('navigates to add content page', async () => {
-      const user = userEvent.setup()
-      render(
-        <SessionProvider>
-          <LibraryPageClient 
-            initialContent={mockContent}
-            initialTotal={mockContent.length}
-            initialPage={1}
-            pageSize={20}
-          />
-        </SessionProvider>
-      )
-
-      // Find the specific "Add Content" button (not "Add Song")
-      const addContentButton = screen.getByRole('button', { name: /add content/i })
-      
-      if (addContentButton) {
-        await user.click(addContentButton)
-        // Should navigate to add content page
-        expect(mockRouterPush).toHaveBeenCalledWith('/add-content')
-      }
-    })
-  })
-})
